// src/model/leaderboards.ts
export interface Leaderboards {
  season: number;
  totalMatches: {
    total: number;
    perDay: number;
    days: number;
  };
  places: Array<{
    category: "wins" | "score" | "matches" | "winRate" | "rating";
    team: string;
    value: number;
    place: number;
  }>;
  playerPlaces: Array<{
    category: "wins" | "score" | "matches" | "winRate" | "rating";
    player: string;
    value: number;
    place: number;
  }>;
}